+++ functions/Set-PSCalendarConfiguration.ps1
@@ -9,14 +9,16 @@ function Set-PSCalendarConfiguration {
         [ValidatePattern('\e\[((?:[0-4]|[39]|[49]|3[0-9]|4[0-9]|38;5;\d+|48;5;\d+);?)+m')]
         [string]$Today,
         [ValidatePattern('\e\[((?:[0-4]|[39]|[49]|3[0-9]|4[0-9]|38;5;\d+|48;5;\d+);?)+m')]
-        [string]$Highlight
+        [string]$Highlight,
+        [ValidatePattern('\e\[((?:[0-4]|[39]|[49]|3[0-9]|4[0-9]|38;5;\d+|48;5;\d+);?)+m')]
+        [string]$Weekend
     )
 
     Write-Verbose "Starting: $($MyInvocation.MyCommand) [v$modVer]"
     Write-Verbose "Using PowerShell version: $($PSVersionTable.PSVersion)"
     Write-Verbose "Running in PowerShell host: $($host.name)"
 
-    $Settings = "Title", "DayOfWeek", "Today", "Highlight"
+    $Settings = "Title", "DayOfWeek", "Today", "Highlight","Weekend"
     foreach ($setting in $Settings) {
         if ($PSBoundParameters.ContainsKey($setting) -AND $PSCmdlet.ShouldProcess($($Setting))) {
             Write-Verbose "Configuring $Setting"